fix wrong input for font width

+++ GoB.py
@@ -1368,7 +1368,7 @@ class GOB_OT_Popup(Operator):
     def invoke(self, context, event):       
         wm = context.window_manager
         font_size_correction = bpy.context.preferences.ui_styles[0].widget_label.points / 10
-        return wm.invoke_props_dialog(self, width = 200 * font_size_correction)
+        return wm.invoke_props_dialog(self, width = int(200 * font_size_correction))
 
     def execute(self, context):
         self.open_addon_prefs(context)
